move gsn to seperate package
